Med Closet Walls & Tileables
There's probably some fancy export process for officially tileable textures, but all the tutorials were from 2016-2017 and long soooo what I did was create a plane that filled the entire (1,1) UV square and filled that with the given texture in Substance Painter. It exported with AO, and I decided not to delete it in case you want a file for messing around in photoshop, but it's otherwise pointless.